Split off configuration into Cabal package.
This is to allow multi-file development of my config.

Began work on a pretty-printer framework for XMonad to interact with
dzen2 (and presumably more), located in
XMonad.Hooks.DynamicLog.PrettyPrinter.DZen2.  It uses the package
`pretty' to acheive its goals, making it much more flexible than the
`PP' facilities inside the DynamicLog package.  This will eventually
be spun off into its own package, likely.

Stole Battery.hs from TaffyBar.  TaffyBar seems not to be up to date
with the latest XMonad, so I guess I won't be trying it any time soon.
Still, the battery stuff seems usable and independant of the rest of
the package.  I should be able to learn stuff about interacting with
DBus as well, since that's how they got the UPower information.