Migrated everything over to StatusText.
Everything compiles and whatever.  Don't try to put stuff in the centre
of the bar or the left side though.  Currently trying to get someone to
release their xfgfont string length calculating code under a BSD/MIT
license, so that I can port it to this and fix the alignment.

Status.Bars contains the functions for actually rendering your
StatusTexts in a nice way.  You create three StatusTextSections which
correspond to the left, centre and right of the bar.  You then pass
these to makeStatusBar and you're basically done.  You then just call
your favourite statusBar function (defaultStatusBar is a fan favourite),
and voila.  A Data.Text with everything you want in it.